Refactor: Extract getCrossPoint to utils.js for better reusability
