+++ s4/scripts/imageProcessing.js
@@ -1,3 +1,5 @@
+import { getCrossPoint } from './utils.js';
+
 var fileData;
 function filer(fileEvent) {
   fileData = fileEvent.target.files[0];
@@ -95,51 +97,34 @@ function drawCanvasSeparater() {
 
 var canvas1, polygon1;
 function drawCanvasClip() {
-  var cpr = getCrossPoint(clipline.aCoords, clipimg.aCoords);
-  var base_size = 1000;
+  const cpr = getCrossPoint(clipline.aCoords, clipimg.aCoords);
+  const base_size = BASE_SIZE;
 
-  var fimg = new fabric.Image(loadedImage.ss);
+  const fimg = new fabric.Image(loadedImage.ss);
   fimg.scaleToWidth(base_size);
   fimg.setControlsVisibility({
-    ml:false, mr:false, mb:false, mt:false,
+    ml: false, mr: false, mb: false, mt: false,
   });
-  var width = fimg.getScaledWidth();
-  var height = fimg.getScaledHeight();
+  const width = fimg.getScaledWidth();
+  const height = fimg.getScaledHeight();
 
-  var topleft = {x: 0, y: 0};
-  var topright = {x: width, y: 0};
-  var bottomright = {x: width, y: height};
-  var bottomleft = {x: 0, y: height};
-  var p = [topleft, topright, bottomright, bottomleft];
+  const topleft = { x: 0, y: 0 };
+  const topright = { x: width, y: 0 };
+  const bottomright = { x: width, y: height };
+  const bottomleft = { x: 0, y: height };
+  const p = [topleft, topright, bottomright, bottomleft];
   if (cpr.top && cpr.bottom) {
     p[1].x = width * cpr.top;
     p[2].x = width * cpr.bottom;
   } else if (cpr.left && cpr.right) {
     p[2].y = height * cpr.right;
     p[3].y = height * cpr.left;
-  } else if (cpr.top && cpr.left) {
-    p[1].x = 0;
-    p[2].x = width * cpr.top;
-    p[2].y = 0;
-    p[3].y = height * cpr.left;
-  } else if (cpr.top && cpr.right) {
-    p[1].x = width * cpr.top;
-    p[2].y = height * cpr.right;
-    p[3].x = width;
-    p[4] = {x: 0, y: height};
-  } else if (cpr.bottom && cpr.left) {
-    p[3].x = width * cpr.bottom;
-    p[4] = {x: 0, y:height * cpr.left};
-  } else if (cpr.bottom && cpr.right) {
-    p[2].y = height * cpr.right;
-    p[3].x = width * cpr.bottom;
-    p[4] = {x: 0, y: height};
   }
 
   if (!canvas1) {
-    var canv = setupCanvas('ss');
-    canvas1 = new fabric.Canvas('ssCanvas', {selection: false, preserveObjectStacking: !true});
-    canvas1.setDimensions({width: width, height: height});
+    const canv = setupCanvas('ss');
+    canvas1 = new fabric.Canvas('ssCanvas', { selection: false, preserveObjectStacking: true });
+    canvas1.setDimensions({ width, height });
     canvas1.controlsAboveOverlay = true;
     canvas1.add(fimg);
   }
@@ -153,30 +138,36 @@ function drawCanvasClip() {
     transparentCorners: false,
     cornerColor: 'blue',
     selectable: false,
-    evented: false
+    evented: false,
   });
   canvas1.clipPath = polygon1;
   canvas1.add(polygon1);
   canvas1.renderAll();
 
   if (!$('#ss').find('input').length) {
-    $('<input>').addClass('rad-button static small dark flat').css({margin:'15px'}).attr('type', 'file').appendTo($('#ss')).change((e) => {
-      filer(e)
-        .then((imgSrc) => loadImage('ss2', imgSrc))
-        .then(() => {
-          var new_img = new Image();
-          loadedImage.step1 = new_img;
-          new_img.onload = () => drawCanvasRemain();
-          new_img.src = canvas1.toDataURL({
-            format: 'png',
-            left: polygon1.left,
-            top: polygon1.top,
-            width: polygon1.getScaledWidth(),
-            height: polygon1.getScaledHeight()
-          });
-        })
-        .catch((msg) => alert(msg));
-    }).appendTo($('#ss'));
+    $('<input>')
+      .addClass('rad-button static small dark flat')
+      .css({ margin: '15px' })
+      .attr('type', 'file')
+      .appendTo($('#ss'))
+      .change((e) => {
+        filer(e)
+          .then((imgSrc) => loadImage('ss2', imgSrc))
+          .then(() => {
+            const new_img = new Image();
+            loadedImage.step1 = new_img;
+            new_img.onload = () => drawCanvasRemain();
+            new_img.src = canvas1.toDataURL({
+              format: 'png',
+              left: polygon1.left,
+              top: polygon1.top,
+              width: polygon1.getScaledWidth(),
+              height: polygon1.getScaledHeight(),
+            });
+          })
+          .catch((msg) => alert(msg));
+      })
+      .appendTo($('#ss'));
   }
 }
 
@@ -278,57 +269,4 @@ function drawCanvasComposite() {
  var canvas = new fabric.Canvas('comCanvas', {selection: false, preserveObjectStacking: !true});
  canvas.setDimensions({width:fimg2.getScaledWidth(), height:fimg2.getScaledHeight()});
  canvas.add(fimg).add(fimg2).renderAll();
-}
-
-function getCrossPoint(lineACoords, imgACoords) {
-  var cp = {};
-  var cpr = {};
-  var pointA = [Math.floor(lineACoords.tl.x), Math.floor(lineACoords.tl.y)];
-  var pointB = [Math.floor(lineACoords.bl.x), Math.floor(lineACoords.bl.y)];
-  var imgSideList = [['tl','tr','top'], ['tr','br','right'], ['br','bl','bottom'], ['bl','tl','left']];
-  for (var e of imgSideList) {
-    var pointC = [Math.floor(imgACoords[e[0]].x), Math.floor(imgACoords[e[0]].y)];
-    var pointD = [Math.floor(imgACoords[e[1]].x), Math.floor(imgACoords[e[1]].y)];
-    var cross_point = _getCrossPoint(pointA, pointB, pointC, pointD);
-    if (cross_point.length > 0) {
-      if (e[2] == 'top' && (cross_point[0] < pointC[0] || cross_point[0] > pointD[0])) cross_point = [];
-      if (e[2] == 'top' && (cross_point[1] < pointC[1] || cross_point[1] > pointD[1])) cross_point = [];
-      if (e[2] == 'right' && (cross_point[0] < pointC[0] || cross_point[0] > pointD[0])) cross_point = [];
-      if (e[2] == 'right' && (cross_point[1] < pointC[1] || cross_point[1] > pointD[1])) cross_point = [];
-      if (e[2] == 'bottom' && (cross_point[0] < pointD[0] || cross_point[0] > pointC[0])) cross_point = [];
-      if (e[2] == 'bottom' && (cross_point[1] < pointD[1] || cross_point[1] > pointC[1])) cross_point = [];
-      if (e[2] == 'left' && (cross_point[0] < pointD[0] || cross_point[0] > pointC[0])) cross_point = [];
-      if (e[2] == 'left' && (cross_point[1] < pointD[1] || cross_point[1] > pointC[1])) cross_point = [];
-      if (cross_point.length && (e[2] == 'top' || e[2] == 'bottom')) cross_point[1] = pointC[1];
-      if (cross_point.length && (e[2] == 'right' || e[2] == 'left')) cross_point[0] = pointC[0];
-    }
-    cp[e[2]] = cross_point;
-    if (cross_point.length > 0) {
-      if (e[2] == 'top') cpr[e[2]] = Math.floor((cross_point[0] - pointC[0]) / (pointD[0] - pointC[0]) * 100) / 100;
-      if (e[2] == 'bottom') cpr[e[2]] = Math.floor((cross_point[0] - pointD[0]) / (pointC[0] - pointD[0]) * 100) / 100;
-      if (e[2] == 'right') cpr[e[2]] = Math.floor((cross_point[1] - pointC[1]) / (pointD[1] - pointC[1]) * 100) / 100;
-      if (e[2] == 'left') cpr[e[2]] = Math.floor((cross_point[1] - pointD[1]) / (pointC[1] - pointD[1]) * 100) / 100;
-    }
-  }
-  return cpr;
-}
-
-function _getCrossPoint(pointA, pointB, pointC, pointD) {
-  pointA = [Math.floor(pointA[0]),Math.floor(pointA[1])];
-  pointB = [Math.floor(pointB[0]),Math.floor(pointB[1])];
-  pointC = [Math.floor(pointC[0]),Math.floor(pointC[1])];
-  pointD = [Math.floor(pointD[0]),Math.floor(pointD[1])];
-  var cross_point = [];
-  var bunbo = (pointB[0] - pointA[0]) * (pointD[1] - pointC[1]) - (pointB[1] - pointA[1]) * (pointD[0] - pointC[0]);
-
-  if (bunbo == 0) return cross_point;
-
-  var vectorAC = [(pointC[0] - pointA[0]), (pointC[1] - pointA[1])];
-  var r = ((pointD[1] - pointC[1]) * vectorAC[0] - (pointD[0] - pointC[0]) * vectorAC[1]) / bunbo;
-  var s = ((pointB[1] - pointA[1]) * vectorAC[0] - (pointB[0] - pointA[0]) * vectorAC[1]) / bunbo;
-
-  var distance = [(pointB[0] - pointA[0]) * r, (pointB[1] - pointA[1]) * r];
-  cross_point = [parseInt(pointA[0] + distance[0]), parseInt(pointA[1] + distance[1])];
-
-  return cross_point;
 }
